# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "StudyLife.Server | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/mcp-connect": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpConnectRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/McpConnectRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/McpConnectRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/McpConnectResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpConnectResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpConnectResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/mcp-assertion-exchange": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpAssertionExchangeRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/McpAssertionExchangeRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/McpAssertionExchangeRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/McpAssertionExchangeResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpAssertionExchangeResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpAssertionExchangeResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/capture-connect": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureConnectRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureConnectRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureConnectRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureConnectResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureConnectResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureConnectResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/capture-assertion-exchange": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureAssertionExchangeRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureAssertionExchangeRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureAssertionExchangeRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureAssertionExchangeResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureAssertionExchangeResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureAssertionExchangeResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/focusguard-connect": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FocusGuardConnectRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FocusGuardConnectRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FocusGuardConnectRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FocusGuardConnectResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusGuardConnectResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusGuardConnectResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/focusguard-assertion-exchange": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FocusGuardAssertionExchangeRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FocusGuardAssertionExchangeRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FocusGuardAssertionExchangeRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FocusGuardAssertionExchangeResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusGuardAssertionExchangeResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusGuardAssertionExchangeResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/focustunes-connect": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FocusTunesConnectRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FocusTunesConnectRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FocusTunesConnectRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FocusTunesConnectResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusTunesConnectResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusTunesConnectResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/focustunes-assertion-exchange": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FocusTunesAssertionExchangeRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FocusTunesAssertionExchangeRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FocusTunesAssertionExchangeRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FocusTunesAssertionExchangeResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusTunesAssertionExchangeResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusTunesAssertionExchangeResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/tray-connect": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrayConnectRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrayConnectRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrayConnectRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrayConnectResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrayConnectResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrayConnectResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/tray-assertion-exchange": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrayAssertionExchangeRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrayAssertionExchangeRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrayAssertionExchangeRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrayAssertionExchangeResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrayAssertionExchangeResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrayAssertionExchangeResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaptureAssertionExchangeRequestDto": {
        "required": [
          "assertion"
        ],
        "type": "object",
        "properties": {
          "assertion": {
            "type": "string"
          }
        }
      },
      "CaptureAssertionExchangeResponseDto": {
        "required": [
          "userId",
          "captureApiKey"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "captureApiKey": {
            "type": "string"
          }
        }
      },
      "CaptureConnectRequestDto": {
        "required": [
          "redirectUri",
          "state"
        ],
        "type": "object",
        "properties": {
          "redirectUri": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        }
      },
      "CaptureConnectResponseDto": {
        "required": [
          "redirectTo"
        ],
        "type": "object",
        "properties": {
          "redirectTo": {
            "type": "string"
          }
        }
      },
...
      "FocusGuardAssertionExchangeRequestDto": {
        "required": [
          "assertion"
        ],
        "type": "object",
        "properties": {
          "assertion": {
            "type": "string"
          }
        }
      },
      "FocusGuardAssertionExchangeResponseDto": {
        "required": [
          "userId",
          "focusGuardApiKey"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "focusGuardApiKey": {
            "type": "string"
          }
        }
      },
      "FocusGuardConnectRequestDto": {
        "required": [
          "redirectUri",
          "state"
        ],
        "type": "object",
        "properties": {
          "redirectUri": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        }
      },
      "FocusGuardConnectResponseDto": {
        "required": [
          "redirectTo"
        ],
        "type": "object",
        "properties": {
          "redirectTo": {
            "type": "string"
          }
        }
      },
...
      "FocusTunesConnectRequestDto": {
        "required": [
          "redirectUri",
          "state"
        ],
        "type": "object",
        "properties": {
          "redirectUri": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        }
      },
      "FocusTunesConnectResponseDto": {
        "required": [
          "redirectTo"
        ],
        "type": "object",
        "properties": {
          "redirectTo": {
            "type": "string"
          }
        }
      },
...
      "McpAssertionExchangeRequestDto": {
        "required": [
          "assertion"
        ],
        "type": "object",
        "properties": {
          "assertion": {
            "type": "string"
          }
        }
      },
      "McpAssertionExchangeResponseDto": {
        "required": [
          "userId",
          "mcpApiKey"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mcpApiKey": {
            "type": "string"
          }
        }
      },
      "McpConnectRequestDto": {
        "required": [
          "redirectUri",
          "state"
        ],
        "type": "object",
        "properties": {
          "redirectUri": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        }
      },
      "McpConnectResponseDto": {
        "required": [
          "redirectTo"
        ],
        "type": "object",
        "properties": {
          "redirectTo": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "SessionToken": {
        "type": "apiKey",
        "description": "Passkey session token issued by POST /api/auth/passkey/complete or the device-link/handoff flows - what the browser/native app clients use.",
        "name": "X-Session-Token",
        "in": "header"
      },
      "ApiKey": {
        "type": "apiKey",
        "description": "Long-lived per-user API key (generated on the Setup page, one slot each for Home Assistant/AI/MCP/Capture - see ApiKeyScopes). Scoped per slot to a subset of endpoints; an out-of-scope request answers 403, not 401.",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
...
}